Split encoding code from kpro_batch to kpro_batch_encoder
Prior to this commit, producers have to collect messages in an
intermediate batch (such as a list or a queue) then call batch encode
when the batch is full.
This commit is the first step to make it possible for a producer to
collect messages into a stateful batch encoder and query the size
(count and bytes) in an API, then send it over a socket immediately
when the batch is full.